HIVE-28613: OTEL: Collect & expose JVM details for LLAP daemons.
