+++ service/src/java/org/apache/hive/service/server/HiveServer2.java
@@ -136,6 +136,8 @@ import org.apache.zookeeper.Watcher;
 import org.apache.zookeeper.ZooDefs.Ids;
 import org.apache.zookeeper.ZooDefs.Perms;
 import org.apache.zookeeper.data.ACL;
+
+import io.opentelemetry.api.GlobalOpenTelemetry;
 import org.eclipse.jetty.servlet.FilterHolder;
 import org.eclipse.jetty.servlet.ServletHolder;
 
@@ -513,8 +515,8 @@ public class HiveServer2 extends CompositeService {
     long otelExporterFrequency =
         hiveConf.getTimeVar(ConfVars.HIVE_OTEL_METRICS_FREQUENCY_SECONDS, TimeUnit.MILLISECONDS);
     if (otelExporterFrequency > 0) {
-      otelExporter = new OTELExporter(AutoConfiguredOpenTelemetrySdk.initialize().getOpenTelemetrySdk(),
-          cliService.getSessionManager(), otelExporterFrequency);
+      otelExporter = new OTELExporter(GlobalOpenTelemetry.get(), cliService.getSessionManager(),
+          otelExporterFrequency, getServerHost());
 
       otelExporter.setName("OTEL Exporter");
       otelExporter.setDaemon(true);
